+++ yshop-mall/src/main/java/co/yixiang/modules/order/vo/ComputeVo.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * @ClassName ComputeVo
@@ -17,10 +18,10 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ComputeVo implements Serializable {
-    private Double couponPrice;
-    private Double deductionPrice;
-    private Double payPostage;
-    private Double payPrice;
-    private Double totalPrice;
+    private BigDecimal couponPrice;
+    private BigDecimal deductionPrice;
+    private BigDecimal payPostage;
+    private BigDecimal payPrice;
+    private BigDecimal totalPrice;
     private Double usedIntegral; //使用了多少积分
 }
